example: move MarkerHighlighter to the markdown example
Move the `DocumentRangeSemanticTokensProvider` demo out of the (already
LSP-heavy) editor example into the markdown example, which is a
`code_editor(Language::Markdown)` and a more natural fit. Add TODO/FIXME/
XXX/HACK/NOTE markers to the markdown fixture as test cases so the provider
has something to highlight on first paint.
