+++ crates/story/examples/markdown.rs
@@ -1,3 +1,6 @@
+use std::ops::Range;
+use std::rc::Rc;
+
 use gpui::{prelude::FluentBuilder as _, *};
 use gpui_component::{
     ActiveTheme as _, IconName, Sizable as _,
@@ -5,12 +8,95 @@ use gpui_component::{
     clipboard::Clipboard,
     h_flex,
     highlighter::Language,
-    input::{Input, InputEvent, InputState, TabSize},
+    input::{
+        DocumentRangeSemanticTokensProvider, Input, InputEvent, InputState, Rope, RopeExt, TabSize,
+    },
     resizable::{h_resizable, resizable_panel},
     text::markdown,
 };
 use gpui_component_assets::Assets;
 use gpui_component_story::Open;
+use lsp_types::{SemanticToken, SemanticTokenType, SemanticTokens, SemanticTokensLegend};
+
+/// Example [`DocumentRangeSemanticTokensProvider`]: tags `TODO` / `FIXME` /
+/// `XXX` / `HACK` / `NOTE` markers anywhere in the document with a
+/// `keyword.special` semantic token so they stand out.
+///
+/// Installed on `input_state.lsp.semantic_tokens_provider`, exactly like the
+/// other LSP providers (`document_color_provider`, `hover_provider`, …). The
+/// editor fetches it (debounced) on document change, caches the result, and
+/// composes it into the render pipeline on top of the tree-sitter syntax
+/// highlighting. This example scans synchronously and returns a ready task;
+/// a real language server would return tokens from an async request, and a
+/// heavy local parser (syntect, …) would offload to a background task.
+struct MarkerHighlighter;
+
+impl MarkerHighlighter {
+    /// Token-type name emitted for every marker. Resolved against the active
+    /// `HighlightTheme`; `keyword.special` falls back to `keyword`.
+    const TOKEN_TYPE: &'static str = "keyword.special";
+}
+
+impl DocumentRangeSemanticTokensProvider for MarkerHighlighter {
+    fn legend(&self) -> SemanticTokensLegend {
+        SemanticTokensLegend {
+            token_types: vec![SemanticTokenType::from(Self::TOKEN_TYPE.to_string())],
+            token_modifiers: vec![],
+        }
+    }
+
+    fn semantic_tokens(
+        &self,
+        text: &Rope,
+        range: Range<usize>,
+        _window: &mut Window,
+        _cx: &mut App,
+    ) -> Task<Result<SemanticTokens>> {
+        const MARKERS: &[&str] = &["TODO", "FIXME", "XXX", "HACK", "NOTE"];
+
+        // Scan the requested range and collect absolute (line, character,
+        // length) hits.
+        let slice = text.slice(range.clone()).to_string();
+        let mut hits: Vec<(u32, u32, u32)> = Vec::new();
+        for marker in MARKERS {
+            let mut from = 0;
+            while let Some(rel) = slice[from..].find(marker) {
+                let abs = range.start + from + rel;
+                let pos = text.offset_to_position(abs);
+                hits.push((pos.line, pos.character, marker.chars().count() as u32));
+                from += rel + marker.len();
+            }
+        }
+        hits.sort_unstable();
+
+        // Delta-encode into LSP semantic tokens — the exact format a real
+        // language server returns from `textDocument/semanticTokens/range`.
+        let mut data = Vec::with_capacity(hits.len());
+        let (mut prev_line, mut prev_char) = (0u32, 0u32);
+        for (line, character, length) in hits {
+            let delta_line = line - prev_line;
+            let delta_start = if delta_line == 0 {
+                character - prev_char
+            } else {
+                character
+            };
+            data.push(SemanticToken {
+                delta_line,
+                delta_start,
+                length,
+                token_type: 0,
+                token_modifiers_bitset: 0,
+            });
+            prev_line = line;
+            prev_char = character;
+        }
+
+        Task::ready(Ok(SemanticTokens {
+            result_id: None,
+            data,
+        }))
+    }
+}
 
 pub struct Example {
     input_state: Entity<InputState>,
@@ -22,7 +108,7 @@ const EXAMPLE: &str = include_str!("./fixtures/test.md");
 impl Example {
     pub fn new(window: &mut Window, cx: &mut Context<Self>) -> Self {
         let input_state = cx.new(|cx| {
-            InputState::new(window, cx)
+            let mut input_state = InputState::new(window, cx)
                 .code_editor(Language::Markdown)
                 .line_number(true)
                 .tab_size(TabSize {
@@ -31,7 +117,13 @@ impl Example {
                 })
                 .searchable(true)
                 .placeholder("Enter your Markdown here...")
-                .default_value(EXAMPLE)
+                .default_value(EXAMPLE);
+
+            // Install the example range semantic tokens provider, alongside
+            // the other LSP providers. It highlights TODO/FIXME/… markers.
+            input_state.lsp.semantic_tokens_provider = Some(Rc::new(MarkerHighlighter));
+
+            input_state
         });
 
         // Focus the input on startup so that actions (e.g. Open) can bubble
